fix(Client): fix size_t bug
do

+++ Client.cpp
@@ -60,7 +60,7 @@ void Client::removeFromServer()
 
 Client* findClient(int fd, std::vector<Client*>& clients)
 {
-	for (int i = 0; i < clients.size(); i++)
+	for (size_t i = 0; i < clients.size(); i++)
 	{
 		if (fd == clients[i]->getFd())
 			return (clients[i]);
